+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 This file is used to list changes made in each version of the firewall cookbook.
 
+## Unreleased
+
 ## 5.0.0 - *2022-04-20*
 
 - Add support for nftables