+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 This file is used to list changes made in each version of the firewall cookbook.
 
-## Unreleased
+## 6.0.0 - *2022-05-09*
 
 - Values for firewalld resources must be specified as one would
   specify them to `firewall-cmd`.